requirements.txt
# AstroAgent v2 — Python 3.11+
# pip install -r requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=23.2.0
httpx>=0.27.0

langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0

hydra-core>=1.3.2
omegaconf>=2.3.0

pydantic>=2.7.0
python-dotenv>=1.0.0
colorlog>=6.8.0

ephem>=4.1.5
astropy>=6.0.0
